preprocess_images/requirements.txt
boto3==1.24.85
pandas==1.4.2
Pillow==9.3.0

